refactor(cli): fold injectable logger into term-log (drop cli-logger)
term-log.mjs is already the human-output surface, so the separate cli-logger
module was a redundant flattened re-wrap of it. Move noopLogger + termLogger
(+ the CliLogger typedef) into term-log.mjs and delete cli-logger.mjs; repoint
the two upgrade importers. Existing `p.log`/`intro`/`outro` consumers are
untouched. Pure internal refactor — no behavior change (upgrade suites 23/23,
strict + eslint clean).
